+++ src/plugins/index.ts
@@ -5,3 +5,12 @@
  */
 
 export { PluginManager, validatePlugin, NAMESPACE_SEPARATOR } from './PluginManager';
+
+// Built-in plugins
+export {
+  datavizPlugin,
+  atmosphericPlugin,
+  scifiPlugin,
+  organicPlugin,
+  corePlugin,
+} from './builtin';